HHH-18868 Cleanup tests and mapper-supers identifier mapper check
