+++ hibernate-core/src/test/java/org/hibernate/orm/test/idclass/mappedsuperclass/MappedSuperclassIdClassAttributesTest.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.orm.test.idclass;
+package org.hibernate.orm.test.idclass.mappedsuperclass;
 
 import jakarta.persistence.metamodel.SingularAttribute;
 import org.hibernate.testing.orm.junit.DomainModel;
@@ -11,14 +11,12 @@ import org.hibernate.testing.orm.junit.SessionFactory;
 import org.hibernate.testing.orm.junit.SessionFactoryScope;
 import org.junit.jupiter.api.Test;
 
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DomainModel(annotatedClasses = {Summary.class, BaseSummary.class})
 @SessionFactory
-@JiraKey( "HHH-18858" )
+@JiraKey("HHH-18858")
 public class MappedSuperclassIdClassAttributesTest {
-
 	@Test
 	public void test(SessionFactoryScope scope) {
 		scope.inSession( entityManager -> {
