+++ hibernate-core/src/test/java/org/hibernate/orm/test/idclass/mappedsuperclass/PKey.java
@@ -2,25 +2,15 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.orm.test.idclass;
-
-import jakarta.persistence.Id;
-import jakarta.persistence.IdClass;
-import jakarta.persistence.MappedSuperclass;
+package org.hibernate.orm.test.idclass.mappedsuperclass;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Objects;
 
-@IdClass(PKey.class)
-@MappedSuperclass
-public class BaseSummary implements Serializable {
+public class PKey implements Serializable {
 
-	@Id
 	private Integer year;
-	@Id
 	private Integer month;
-	private BigDecimal value;
 
 	public Integer getYear() {
 		return year;
@@ -38,21 +28,13 @@ public class BaseSummary implements Serializable {
 		this.month = month;
 	}
 
-	public BigDecimal getValue() {
-		return value;
-	}
-
-	public void setValue(BigDecimal value) {
-		this.value = value;
-	}
-
 	@Override
 	public boolean equals(Object o) {
 		if ( o == null || getClass() != o.getClass() ) {
 			return false;
 		}
-		BaseSummary that = (BaseSummary) o;
-		return Objects.equals( year, that.year ) && Objects.equals( month, that.month );
+		PKey pKey = (PKey) o;
+		return Objects.equals( year, pKey.year ) && Objects.equals( month, pKey.month );
 	}
 
 	@Override
