+++ hibernate-core/src/test/java/org/hibernate/orm/test/idclass/mappedsuperclass/Summary.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.orm.test.idclass;
+package org.hibernate.orm.test.idclass.mappedsuperclass;
 
 import jakarta.persistence.Entity;
